fix: charts tooltips positioning
